requirements.txt
requests==2.12.4
pyxdg==0.25
